backend/requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
psycopg2-binary==2.8.3
mysqlclient==2.0.1
djangorestframework==3.11.1
pyyaml==5.3.1
django-otp==1.0.0
qrcode==6.1
